- add validation

+++ app/src/main/java/com/harris/cryptoworld/presentation/cryptodetail/CryptoDetailScreen.kt
@@ -59,7 +59,7 @@ fun CryptoDetailScreen(viewModel: CryptoViewModel) {
 
                         value = enteredAmount.toString(),
                         onValueChange = { query ->
-                            viewModel.onAmountChange(query.toDouble(), conversionRate)
+                            viewModel.onAmountChange(query, conversionRate)
                         },
                         keyboardOptions = KeyboardOptions(
                             keyboardType = KeyboardType.Number
